requirements-training.txt
webdataset>=0.2.5
pandas
tqdm
einops
numpy
tensorboard
wandb
open_clip_torch
embedding_reader
